+++ src/Fauna/Output/text_table_writer.h
@@ -136,10 +136,13 @@ class TextTableWriter : public WriterInterface {
 
   /** @{ \name File Streams */
   std::ofstream available_forage;
+  std::ofstream body_fat;
   std::ofstream digestibility;
   std::ofstream eaten_forage_per_ind;
   std::ofstream eaten_nitrogen_per_ind;
-  std::ofstream mass_density_per_hft;
+  std::ofstream individual_density;
+  std::ofstream mass_density;
+  std::ofstream mass_density_per_hft;  // deprecated
   // Add new output variables here (alphabetical order).
   /** @} */  // File Streams
 };