+++ CHANGELOG.md
@@ -18,6 +18,10 @@ The format is based on [Keep a Changelog][] by Olivier Lacan, and this project a
 - Code coverage badge with [codecov.io](https://codecov.io/gh/wtraylor/modular_megafauna_model/)
 - Software metadata in [codemeta format](https://codemeta.github.io/)
 - Add parameter sanity checks: mortality must not exceed reproduction; expenditure must not exceed intake [#45]
+- New text table output: `individual_density` and `body_fat` [#43]
+
+### Changed
+- Rename output file `mass_density_per_hft` to `mass_density` [#44]
 
 ### Fixed
 - When configuring CMake with `BUILD_DEMO_SIMULATOR=ON` but `BUILD_TESTING=OFF`, the `megafauna.toml` instruction file was not copied even though `run_demo` needs it.
@@ -266,4 +270,6 @@ The format is based on [Keep a Changelog][] by Olivier Lacan, and this project a
 [#29]: https://github.com/wtraylor/modular_megafauna_model/issues/29
 [#31]: https://github.com/wtraylor/modular_megafauna_model/issues/31
 [#32]: https://github.com/wtraylor/modular_megafauna_model/issues/32
+[#43]: https://github.com/wtraylor/modular_megafauna_model/issues/43
+[#44]: https://github.com/wtraylor/modular_megafauna_model/issues/44
 [#45]: https://github.com/wtraylor/modular_megafauna_model/issues/45